fix animations, swipe etc.
